// components/brave_wallet_ui/constants/types.ts
export type SafeBlowfishWarning = {
  severity: BraveWallet.BlowfishWarningSeverity
  kind: BraveWallet.BlowfishWarningKind
  /**
   * human-readable message to present to the end-user
   */
  message: string
}
// ...
interface SafeBlowfishEVMError {
  kind: BraveWallet.BlowfishEVMErrorKind
  humanReadableError: string
}

interface SafeBlowfishSolanaError {
  kind: BraveWallet.BlowfishSolanaErrorKind
  humanReadableError: string
}

export type SafeEVMStateChange<
  UNION_KEY extends //
  keyof BraveWallet.BlowfishEVMStateChangeRawInfoDataUnion //
  = keyof BraveWallet.BlowfishEVMStateChangeRawInfoDataUnion
> = {
  humanReadableDiff: string
  rawInfo: SafeEvmRawInfo<UNION_KEY>
}

export type SafeERC20ApprovalEvent = SafeEVMStateChange<
  'erc20ApprovalData'
>

export type SafeERC721ApprovalEvent = SafeEVMStateChange<
  'erc721ApprovalData'
>

export type SafeERC721ApprovalForAllEvent = SafeEVMStateChange<
  'erc721ApprovalForAllData'
>

export type SafeERC1155ApprovalForAllEvent = SafeEVMStateChange<
  'erc1155ApprovalForAllData'
>

export type SafeERC20TransferEvent = SafeEVMStateChange<
  'erc20TransferData'
>

export type SafeERC721TransferEvent = SafeEVMStateChange<
  'erc721TransferData'
>

export type SafeERC1155TransferEvent = SafeEVMStateChange<
  'erc1155TransferData'
>

export type SafeNativeTransferEvent = SafeEVMStateChange<
  'nativeAssetTransferData'
>

export type SafeEvmTransferEvent =
  | SafeERC20TransferEvent
  | SafeERC721TransferEvent
  | SafeERC1155TransferEvent
  | SafeNativeTransferEvent

export type SafeEvmApprovalEvent =
  | SafeERC20ApprovalEvent
  | SafeERC721ApprovalEvent
  | SafeERC721ApprovalForAllEvent
  | SafeERC1155ApprovalForAllEvent

export type SafeEvmEvent = SafeEvmApprovalEvent | SafeEvmTransferEvent

export type SafeSolanaStateChange<
  UNION_KEY extends //
  keyof BraveWallet.BlowfishSolanaStateChangeRawInfoDataUnion //
  = keyof BraveWallet.BlowfishSolanaStateChangeRawInfoDataUnion
> = {
  humanReadableDiff: string
  rawInfo: SafeSolanaRawInfo<UNION_KEY>
  suggestedColor: BraveWallet.BlowfishSuggestedColor
}

export type SafeSplApprovalEvent = SafeSolanaStateChange<
  'splApprovalData'
>

export type SafeSolanaStakeChangeEvent = SafeSolanaStateChange<
  'solStakeAuthorityChangeData'
>

export type SafeSolanaAccountOwnerChangeEvent = SafeSolanaStateChange<
  'solStakeAuthorityChangeData' // TODO: not implemented in core
>

export type SafeSolTransferEvent = SafeSolanaStateChange<
  'solTransferData'
>

export type SafeSplTransferEvent = SafeSolanaStateChange<
  'splTransferData'
>

type SafeSolanaEvent =
  | SafeSolTransferEvent
  | SafeSplApprovalEvent
  | SafeSplTransferEvent
  | SafeSolanaStakeChangeEvent
  | SafeSolanaAccountOwnerChangeEvent

export interface SafeEVMSimulationResults {
  error: SafeBlowfishEVMError | undefined
  expectedStateChanges: SafeEvmEvent[]
};

export interface SafeSolanaSimulationResults {
  error: SafeBlowfishSolanaError | undefined
  expectedStateChanges: SafeSolanaEvent[]
};

type SafeEvmRawInfo<
  DATA_UNION_KEY //
  extends keyof BraveWallet.BlowfishEVMStateChangeRawInfoDataUnion
> = {
  kind: BraveWallet.BlowfishEVMRawInfoKind
  data: Record<
    DATA_UNION_KEY,
    Exclude<
      BraveWallet.BlowfishEVMStateChangeRawInfoDataUnion[DATA_UNION_KEY],
      undefined
    >
  >
}

type SafeSolanaRawInfo<
  DATA_UNION_KEY //
  extends keyof BraveWallet.BlowfishSolanaStateChangeRawInfoDataUnion
> = {
  kind: BraveWallet.BlowfishSolanaRawInfoKind
  data: Record<
    DATA_UNION_KEY,
    Exclude<
      BraveWallet.BlowfishSolanaStateChangeRawInfoDataUnion[DATA_UNION_KEY],
      undefined
    >
  >
}

type SafeBlowfishResponseBase = {
  action: BraveWallet.BlowfishSuggestedAction
  warnings: SafeBlowfishWarning[]
}

export type SafeBlowfishEvmResponse = SafeBlowfishResponseBase & {
  simulationResults: SafeEVMSimulationResults
}

export type SafeBlowfishSolanaResponse = SafeBlowfishResponseBase & {
  simulationResults: SafeSolanaSimulationResults
}